pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invoiceflow"
version = "0.1.0"
description = "Automatic invoice extraction and organization for freelancers"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "anthropic>=0.25.0",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.3.0",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "gspread>=6.1.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "watchdog>=4.0.0",
    "aiosqlite>=0.20.0",
    "jinja2==3.1.4",
    "starlette==0.41.3",
    "python-dateutil>=2.9.0",
    "httpx>=0.27.0",
    "resend>=2.0.0",
]

[project.scripts]
invoiceflow = "invoiceflow.main:run"

[tool.hatch.build.targets.wheel]
packages = ["invoiceflow"]
